Fixing the aiida-daemon service for aiida 1.0b2
Fixes #8

In particular, now added dependency on rabbitmq and fixed
the PIDFile directive.
Regarding point 3 of the issue (inconsistency btw using
the service and starting/stopping the service manually)
I don't think that, at the moment, there is a risk of
unexpected behavior, so I am not fixing that part.