fix aiida_data_folder issue
Some (default) folder locations contain the `${HOME}` variable - this is
a bit of a hack to worka round an unresolved ansible issue:
{{ ansible_env.HOME }} resolves to the home of the ansible user and is
unaffected by *becoming* a different user.

So, we user the current_user role to store the home location of the user
we want to target, and then replace the `${HOME}` variable in folder
paths.

This was not done for the AiiDA data and template folders, which is now
fixed by this commit.