fix systemd templates for profile name with dashes
When a profile name included dashes (e.g. my-profile), the dashes were
replaced by slashes in the systemd unit files for the AiiDA daemon / the
REST API, resulting in an error upon start (profile "my/profile" not
found).

This PR fixes the unit templates by switching from the systemd unit
instance placeholder %i to placeholder %I ("Same as "%i", but with
escaping undone.").
It also sets a profile name with a dash for the molecule CI tests.